prayon_tracking/requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-crispy-forms==1.8.1
django-import-export==2.0.1
django-pandas==0.6.1
django-tables2==2.2.1
et-xmlfile==1.0.1
img2pdf==0.3.3
jdcal==1.4.1
MarkupPy==1.14
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
jinja2==2.11.1
